refactor: split organizationService into multiple services
