+++ marble-backoffice/src/services/OrganizationService.ts
@@ -1,52 +1,21 @@
-import { useCallback, useEffect, useState } from "react";
-import type {
-  AstNodeEvaluation,
-  DataModel,
-  Iteration,
-  Organization,
-  Scenario,
-  ScenarioValidation,
-} from "@/models";
+import { useCallback, useState } from "react";
+import type { Organization } from "@/models";
 import { PageLink } from "@/models";
 import {
   type OrganizationRepository,
-  type ScenariosRepository,
   type UserRepository,
   fetchAllOrganizations,
   fetchOrganization,
   postOrganization,
-  fetchScenariosOfOrganization,
-  fetchScenario,
   deleteOrganization,
   patchOrganization,
-  fetchDataModelOfOrganization,
-  replaceDataModelOfOrganization,
-  postScenario,
-  fetchIteration,
-  postIteration,
-  patchIteration,
-  postRule,
-  updateRule,
-  publishIteration,
-  validateIteration,
-  fetchIterationsOfScenario,
 } from "@/repositories";
 import { useSimpleLoader } from "@/hooks/SimpleLoader";
 import { showLoader, type LoadingDispatcher } from "@/hooks/Loading";
 import { useNavigate } from "react-router-dom";
-import {
-  adaptDataModelApiResult,
-  adaptDataModelDto,
-} from "@/models/DataModelDto";
-import {
-  type ExampleRule,
-  demoRules,
-  exampleTriggerCondition,
-} from "./ExampleAst";
 
 export interface OrganizationService {
   organizationRepository: OrganizationRepository;
-  scenariosRepository: ScenariosRepository;
   userRepository: UserRepository;
 }
 
@@ -113,316 +82,6 @@ export function useDeleteOrganization(service: OrganizationService) {
   );
 
   return { deleteOrganization: deleteOrg };
-}
-
-export function useScenarios(
-  service: OrganizationService,
-  loadingDispatcher: LoadingDispatcher,
-  organizationId: string
-) {
-  const loadScenarios = useCallback(() => {
-    return fetchScenariosOfOrganization(
-      service.scenariosRepository,
-      organizationId
-    );
-  }, [service, organizationId]);
-
-  const [scenarios, refreshScenarios] = useSimpleLoader<Scenario[]>(
-    loadingDispatcher,
-    loadScenarios
-  );
-
-  return {
-    scenarios,
-    refreshScenarios,
-  };
-}
-
-// function findDraft(iterations: Iteration[]): Iteration | null {
-//   return iterations.find((iteration) => iteration.version === null) ?? null;
-// }
-
-export function useSingleScenario({
-  service,
-  loadingDispatcher,
-  scenarioId,
-  iterationId,
-}: {
-  service: OrganizationService;
-  loadingDispatcher: LoadingDispatcher;
-  scenarioId: string;
-  iterationId: string | null;
-}) {
-  const loadScenario = useCallback(async () => {
-    const scenario = await fetchScenario(
-      service.scenariosRepository,
-      scenarioId
-    );
-
-    scenario.allIterations = await fetchIterationsOfScenario(
-      service.scenariosRepository,
-      scenario.organizationId,
-      scenarioId
-    );
-
-    return scenario;
-  }, [service.scenariosRepository, scenarioId]);
-
-  const [scenario, refreshScenario] = useSimpleLoader<Scenario>(
-    loadingDispatcher,
-    loadScenario
-  );
-
-  const loadIteration = useCallback(async () => {
-    if (scenario === null || iterationId === null) {
-      return null;
-    }
-    return await fetchIteration(
-      service.scenariosRepository,
-      scenario.organizationId,
-      iterationId
-    );
-  }, [iterationId, scenario, service.scenariosRepository]);
-
-  const [iteration, refreshIteration] = useSimpleLoader<Iteration>(
-    loadingDispatcher,
-    loadIteration
-  );
-
-  return {
-    scenario,
-    iteration,
-    refreshScenario: async () => {
-      await refreshScenario();
-      await refreshIteration();
-    },
-  };
-}
-
-function nodeEvaluationErrors(node: AstNodeEvaluation): string[] {
-  return [
-    ...(node.errors ?? []).map((e) => e.message),
-    ...node.children.flatMap(nodeEvaluationErrors),
-    ...Object.values(node.namedChildren).flatMap(nodeEvaluationErrors),
-  ];
-}
-
-function scenarioValidationErrors(validation: ScenarioValidation): string[] {
-  const errors: string[] = [
-    ...validation.errors.map((e) => `Error: ${e}`),
-    ...nodeEvaluationErrors(validation.triggerEvaluation).map(
-      (e) => `Trigger: ${e}`
-    ),
-    ...Object.values(validation.rulesEvaluations)
-      .map(nodeEvaluationErrors)
-      .flat()
-      .map((e: string) => `Rule: ${e}`),
-  ];
-
-  return errors;
-}
-
-export function useAddScenarios(
-  service: OrganizationService,
-  loadingDispatcher: LoadingDispatcher,
-  organizationId: string,
-  refreshScenarios: () => Promise<void>
-) {
-  return {
-    addDemoScenario: async () => {
-      const createDemoScenario = async () => {
-        // Post a new scenario
-        const scenario = await postScenario(
-          service.organizationRepository,
-          organizationId,
-          {
-            name: "Demo scenario",
-            description: "Demo scenario",
-            triggerObjectType: "transactions",
-          }
-        );
-
-        // post a new iteration
-        const iterationId = (
-          await postIteration(
-            service.scenariosRepository,
-            organizationId,
-            scenario.scenarioId
-          )
-        ).iterationId;
-
-        // patch the iteration
-        await patchIteration(
-          service.scenariosRepository,
-          organizationId,
-          iterationId,
-          {
-            scoreReviewThreshold: 20,
-            scoreRejectThreshold: 30,
-            schedule: "*/10 * * * *",
-            triggerCondition: exampleTriggerCondition,
-          }
-        );
-
-        const validation = await validateIteration(
-          service.scenariosRepository,
-          iterationId
-        );
-
-        // check for errors in trigger evaluation
-        const errors = nodeEvaluationErrors(validation.triggerEvaluation);
-        if (errors.length > 0) {
-          throw new Error(errors.join("\n"));
-        }
-
-        const createExampleRule = async (exampleRule: ExampleRule) => {
-          const rule = await postRule(
-            service.scenariosRepository,
-            organizationId,
-            iterationId
-          );
-
-          await updateRule(
-            service.scenariosRepository,
-            organizationId,
-            rule.ruleId,
-            {
-              ...exampleRule,
-            }
-          );
-
-          const validation = await validateIteration(
-            service.scenariosRepository,
-            iterationId
-          );
-
-          const errors = scenarioValidationErrors(validation);
-          if (errors.length > 0) {
-            throw new Error(errors.join("\n"));
-          }
-        };
-
-        // post a rules
-        for (const exampleRule of demoRules) {
-          await createExampleRule(exampleRule);
-        }
-
-        await publishIteration(
-          service.scenariosRepository,
-          organizationId,
-          iterationId
-        );
-
-        await refreshScenarios();
-      };
-
-      await showLoader(loadingDispatcher, createDemoScenario());
-    },
-  };
-}
-
-export function useDataModel(
-  service: OrganizationService,
-  loadingDispatcher: LoadingDispatcher,
-  organizationId: string
-) {
-  const loadDataModel = useCallback(async () => {
-    const dataModel = await fetchDataModelOfOrganization(
-      service.scenariosRepository,
-      organizationId
-    );
-    return dataModel;
-  }, [service, organizationId]);
-
-  const [dataModel] = useSimpleLoader<DataModel>(
-    loadingDispatcher,
-    loadDataModel
-  );
-
-  return {
-    dataModel,
-  };
-}
-
-export function useEditDataModel(
-  service: OrganizationService,
-  loadingDispatcher: LoadingDispatcher,
-  organizationId: string,
-  dataModel: DataModel | null
-) {
-  const [dataModelString, setDataModelString] = useState<string | null>(null);
-  const [dataModelError, setDataModelError] = useState<string>("");
-  const [validatedDataModel, setValidatedDataModel] =
-    useState<DataModel | null>(null);
-
-  const setBackendModel = useCallback((dataModel: DataModel) => {
-    setDataModelString(JSON.stringify(adaptDataModelDto(dataModel), null, 2));
-  }, []);
-
-  useEffect(() => {
-    if (dataModel !== null && dataModelString === null) {
-      setBackendModel(dataModel);
-    }
-  }, [dataModel, dataModelString, setBackendModel]);
-
-  const validateDataModel = useCallback(() => {
-    try {
-      const data = JSON.parse(dataModelString || "");
-      setValidatedDataModel(
-        adaptDataModelApiResult({
-          data_model: data,
-        })
-      );
-      setDataModelError("");
-    } catch (e) {
-      setValidatedDataModel(null);
-      if (e instanceof Error) {
-        setDataModelError(e.message);
-      }
-    }
-  }, [dataModelString]);
-
-  useEffect(() => {
-    validateDataModel();
-  }, [dataModelString, validateDataModel]);
-
-  const [saveDataModelAlertDialogOpen, setSaveDataModelAlertDialogOpen] =
-    useState(false);
-
-  const canSave = validatedDataModel !== null;
-  const saveDataModel = async () => {
-    if (validatedDataModel === null) {
-      return;
-    }
-    setSaveDataModelAlertDialogOpen(true);
-  };
-
-  const saveDataModelConfirmed = async () => {
-    if (validatedDataModel === null) {
-      throw Error("Cannot save invalid data model");
-    }
-    const newDataModel = await showLoader(
-      loadingDispatcher,
-      replaceDataModelOfOrganization(
-        service.organizationRepository,
-        organizationId,
-        validatedDataModel
-      )
-    );
-    setBackendModel(newDataModel);
-    setSaveDataModelAlertDialogOpen(false);
-  };
-
-  return {
-    dataModelString,
-    setDataModelString,
-    saveDataModel,
-    saveDataModelConfirmed,
-    dataModelError,
-    saveDataModelAlertDialogOpen,
-    setSaveDataModelAlertDialogOpen,
-    canSave,
-  };
 }
 
 interface OrganizationViewModel {
