+++ marble-backoffice/src/services/index.ts
@@ -6,3 +6,6 @@ export * from "./DecisionService";
 export * from "./ApiKeyService";
 export * from "./AstExpressionService";
 export * from "./AstEditorService";
+export * from "./ScenarioService";
+export * from "./DataModelService";
+export * from "./DemoScenarioService";
